reorganise tests and package dirs
