+++ mcp_template/utils/__init__.py
@@ -6,3 +6,4 @@ from pathlib import Path
 
 ROOT_DIR = Path(__file__).parent.parent.parent
 TEMPLATES_DIR = ROOT_DIR / "templates"
+TESTS_DIR = ROOT_DIR / "tests"
