+++ tests/test_configuration.py
@@ -5,7 +5,6 @@ Tests for configuration mapping and type conversion functionality.
 import json
 import tempfile
 from pathlib import Path
-from unittest.mock import Mock
 
 import pytest
