+++ tests/test_template/test_all_templates.py
